requirements.txt
requests>=2.28.0
tweepy>=4.0.0
selenium>=4.0.0
beautifulsoup4>=4.0.0
schedule>=1.2.0
mysql-connector-python>=8.0.0
